evoting-app-backend/requirements.txt
django>=4.2,<5.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
